llm-chatbot/llm-orchestration/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
ollama==0.1.3
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
python-dateutil==2.8.2
